Rename spiceai groupid to ai.spice
