+++ src/main/java/ai/spice/Config.java
@@ -20,7 +20,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package spiceai;
+package ai.spice;
 
 import java.net.URI;
 import java.net.URISyntaxException;
