+++ src/test/java/ai/spice/FlightQueryTest.java
@@ -20,7 +20,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package spiceai;
+package ai.spice;
 
 import java.util.concurrent.ExecutionException;
